requirements.txt
# Production requirements for Swarm Multi-Agent System v2.0

# Core Flask framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Database drivers
psycopg2-binary==2.9.10  # PostgreSQL driver
SQLAlchemy==2.0.23

# Authentication and security
PyJWT==2.10.1
bcrypt==4.3.0
cryptography>=40.0.0,<42.0.0

# HTTP client and API integrations
requests==2.31.0
httpx==0.28.1

# Production server
gunicorn==21.2.0

# Development and testing tools
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
coverage==7.3.2

# Code quality tools
black==25.1.0
isort==6.0.1
flake8==7.3.0
mypy==1.16.1
bandit==1.8.5
safety==3.5.2
pre-commit==4.2.0

# Optional: Redis for caching and sessions
redis==5.0.1

# Optional: WebSocket support for real-time features
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# Utilities
python-dotenv==1.0.0
click==8.1.8

